Initial version
For now, this connector is not all that useful because it makes no
attempt to import the crew from the Flight Office booking to the
Wi-Flight reservation, and it makes no attempt to ask Wi-Flight to
notify Flight Office of flight IDs of successful matches. But it
is the skeleton of a working connector which successfully polls
new or changed bookings from Flight Office and imports them to
Wi-Flight.